Move server IT tests under server tests
