+++ src/test/java/org/glassfish/main/distributions/docker/server/AsadminIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025 Contributors to the Eclipse Foundation.
+ * Copyright (c) 2024,2025 Contributors to the Eclipse Foundation.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -13,8 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-
-package org.glassfish.main.distributions.docker;
+package org.glassfish.main.distributions.docker.server;
 
 import java.net.http.HttpResponse;
 
@@ -23,19 +22,22 @@ import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-import static org.glassfish.main.distributions.docker.HttpUtilities.getServerDefaultRoot;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.stringContainsInOrder;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.glassfish.main.distributions.docker.server.HttpUtilities.getServerDefaultRoot;
 
+/**
+ *
+ */
 @Testcontainers
-public class DefaultIT {
+public class AsadminIT {
 
     @SuppressWarnings({"rawtypes", "resource"})
     @Container
     private final GenericContainer server = new GenericContainer<>(System.getProperty("server.docker.glassfish.image"))
-        .withExposedPorts(8080)
-        .withLogConsumer(o -> System.err.print("GF: " + o.getUtf8String()));
+            .withCommand("asadmin start-domain").withExposedPorts(8080)
+            .withLogConsumer(o -> System.err.print("GF: " + o.getUtf8String()));
 
     @Test
     void rootResourceGivesOkWithDefaultResponse() throws Exception {
