+++ src/test/java/org/glassfish/main/distributions/docker/server/UserPassword.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package org.glassfish.main.distributions.docker;
+package org.glassfish.main.distributions.docker.server;
 
 /**
  *
